fix: 파일 위치 변경, 작업 폴더 구성 변경
