+++ css/signin.css
@@ -1,14 +1,4 @@
-:root {
-  --primary-color: #6d6afe;
-  --red-color: #ff5b56;
-  --black-color: #111322;
-  --white-color: #ffffff;
-  --gray1-color: #3e3e43;
-  --gray2-color: #9fa6b2;
-  --gray3-color: #ccd5e3;
-  --gray4-color: #e7effb;
-  --gray5-color: #f0f6ff;
-}
+@import url('../css/color.css');
 
 html {
   font-size: 10px;
