+++ index.html
@@ -11,9 +11,8 @@
     <meta property="twitter:title" content="Linkbrary" />
     <meta property="twitter:description" content="세상의 모든 정보를 쉽게 저장하고 관리해보세요" />
     <meta property="twitter:url" content="https://earnest-smakager-f9a03c.netlify.app/" />
-    <link rel="stylesheet" href="./reset.css" />
-    <link rel="stylesheet" href="./style.css" />
-    <link rel="stylesheet" href="./media.css" />
+    <link rel="stylesheet" href="./css/reset.css" />
+    <link rel="stylesheet" href="./css/main.css" />
     <title>Linkbrary</title>
   </head>
   <body>
@@ -21,11 +20,11 @@
       <ul class="header_wrap">
         <li class="logo">
           <a href="#">
-            <img src="./asset/logo.svg" alt="Linkbrary" />
+            <img src="./images/icon_svg/logo.svg" alt="Linkbrary" />
           </a>
         </li>
         <li class="login">
-          <a href="./signin/">로그인</a>
+          <a href="./pages/signin.html">로그인</a>
         </li>
       </ul>
     </header>
@@ -36,8 +35,8 @@
             <h1><span class="text_gradient info_gradient">세상의 모든 정보</span>를<br /></h1>
             <h1>쉽게 저장하고 관리해 보세요</h1>
           </div>
-          <a href="./signup/" class="link_add">링크 추가하기</a>
-          <img src="./asset/main_header.png" alt="" />
+          <a href="./pages/signup.html" class="link_add">링크 추가하기</a>
+          <img src="./images/main_header.png" alt="" />
         </div>
       </div>
       <section>
@@ -52,7 +51,7 @@
                 한 공간에 저장하세요.
               </p>
             </div>
-            <img src="./asset/main_link.png" alt="" />
+            <img src="./images/main_link.png" alt="" />
           </div>
         </div>
         <div class="card">
@@ -65,7 +64,7 @@
                 다양하게 활용할 수 있습니다.
               </p>
             </div>
-            <img src="./asset/main_folder.png" alt="" />
+            <img src="./images/main_folder.png" alt="" />
           </div>
         </div>
         <div class="card">
@@ -79,7 +78,7 @@
                 공유해 보세요.
               </p>
             </div>
-            <img src="./asset/main_share.png" alt="" />
+            <img src="./images/main_share.png" alt="" />
           </div>
         </div>
         <div class="card">
@@ -89,7 +88,7 @@
               <h1><span class="text_gradient sns_gradient">검색</span>해 보세요</h1>
               <p class="image_description">중요한 정보들을 검색으로 쉽게 찾아보세요.</p>
             </div>
-            <img src="./asset/main_sns.png" alt="" />
+            <img src="./images/main_sns.png" alt="" />
           </div>
         </div>
       </section>
@@ -105,21 +104,20 @@
         </div>
         <div class="sns_fill_wrap">
           <a href="https://www.facebook.com/" target="_blank">
-            <img src="./asset/fill_facebook_icon.svg" alt="facebook" />
+            <img src="./images/icon_svg/fill_facebook_icon.svg" alt="facebook" />
           </a>
           <a href="https://twitter.com/?lang=ko" target="_blank">
-            <img src="./asset/fill_twitter_icon.svg" alt="twitter" />
+            <img src="./images/icon_svg/fill_twitter_icon.svg" alt="twitter" />
           </a>
           <a href="https://www.youtube.com/?gl=KR&hl=ko" target="_blank">
-            <img src="./asset/fill_youtube_icon.svg" alt="youtube" />
+            <img src="./images/icon_svg/fill_youtube_icon.svg" alt="youtube" />
           </a>
           <a href="https://www.instagram.com/" target="_blank">
-            <img src="./asset/fill_instagram_icon.svg" alt="instagram" />
+            <img src="./images/icon_svg/fill_instagram_icon.svg" alt="instagram" />
           </a>
         </div>
       </div>
     </footer>
-    <script type="module" src="./main.js"></script>
-    <script src="./scrollNavShadow.js"></script>
+    <script type="module" src="./script/main.js"></script>
   </body>
 </html>
