--- media.css
@@ -1,59 +1,0 @@
-@media screen and (min-width: 1200px) {
-  header {
-    padding: 2rem 10rem;
-  }
-}
-/* pc */
-
-@media screen and (max-width: 1199px) {
-  header {
-    padding: 2rem 3.2rem;
-  }
-  main .main_header img {
-    width: 100%;
-  }
-  section .card {
-  }
-  section .card .card_item {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  footer .footer_wrap {
-    padding: 3.2rem 10.4rem 10.8rem;
-    display: grid;
-    grid-template-columns: 1fr;
-    grid-template-rows: repeat(3, 1fr);
-    justify-items: center;
-    gap: 2rem;
-    color: #676767;
-  }
-  footer .sns_fill_wrap {
-    justify-self: center;
-  }
-}
-
-/* tablet */
-
-@media screen and (max-width: 767px) {
-  main .main_header h1,
-  main .main_header span {
-    font-size: 3.2rem;
-  }
-
-  main .main_header a.link_add {
-    margin: 2.4rem;
-  }
-  section .card .card_item {
-    flex-direction: column;
-  }
-  section .card {
-    padding: 4rem 3.2rem;
-  }
-  section .card .card_item img {
-    width: 100%;
-  }
-}
-
-/* mobile */
