+++ pages/signin.html
@@ -3,15 +3,17 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="../reset.css" />
-    <link rel="stylesheet" href="./signin.css" />
+    <link rel="stylesheet" href="../css/reset.css" />
+    <link rel="stylesheet" href="../css/signin.css" />
     <title>SignIn</title>
   </head>
   <body>
     <main>
       <header>
-        <a href="../index.html" class="logo"><img src="../asset/logo.svg" alt="Linkbrary" /></a>
-        <p>회원이 아니신가요? <a href="../signup/" class="signup_button">회원 가입하기</a></p>
+        <a href="../index.html" class="logo"
+          ><img src="../images/icon_svg/logo.svg" alt="Linkbrary"
+        /></a>
+        <p>회원이 아니신가요? <a href="./signup.html" class="signup_button">회원 가입하기</a></p>
       </header>
       <section>
         <form class="signin_form" action="">
@@ -21,7 +23,7 @@
             <!-- 비번 표시 토글 버튼 기능을 위한 div-->
             <label for="passwordToggle" class="toggle_block">비밀번호</label>
             <input id="passwordToggle" name="password" type="password" placeholder="******" />
-            <img src="../asset/eye-on.svg" alt="" class="password_eyes" />
+            <img src="../images/icon_svg/eye-on.svg" alt="" class="password_eyes" />
           </div>
           <button type="submit" class="signin_button">로그인</button>
         </form>
@@ -30,14 +32,14 @@
         <p>소셜 로그인</p>
         <div class="sns_wrap">
           <a href="https://www.google.com/">
-            <img src="../asset/google_icon.svg" alt="google" />
+            <img src="../images/icon_svg/google_icon.svg" alt="google" />
           </a>
           <a href="https://www.kakaocorp.com/page/">
-            <img src="../asset/kakao_icon.svg" alt="kakao" />
+            <img src="../images/icon_svg/kakao_icon.svg" alt="kakao" />
           </a>
         </div>
       </footer>
     </main>
-    <script type="module" src="../main.js"></script>
+    <script type="module" src="../script/signin.js"></script>
   </body>
 </html>
