--- utils/scrollCardInner.js
@@ -1,5 +1,0 @@
-const scrollCardInnerEvent = (e) => {
-  console.log(e);
-};
-
-window.addEventListener('scroll', scrollCardInnerEvent);
